Collision, Repetition of Obstacles, Bonus
Added the code for repetition of obstacles and collision, but this still needs to be modified. Bonus was added into the game, but this needs to be modified as well so that it appears on the window. Collision should be modified at the end so that it can be done for both the obstacles and the bonus.